+++ hack/tests/subcommand.sh
@@ -16,32 +16,37 @@ operator-sdk test local ./test/e2e --global-manifest deploy/crds/cache.example.c
 
 # we use the test-memcached namespace for all future tests, so we only need to set this trap once
 kubectl create namespace test-memcached
-trap_add 'kubectl delete namespace test-memcached || true' EXIT
-operator-sdk test local ./test/e2e --namespace=test-memcached
+trap_add 'kubectl delete --ignore-not-found namespace test-memcached' EXIT
+operator-sdk test local ./test/e2e --deploy-namespace=test-memcached
 kubectl delete namespace test-memcached
 
 # test operator in up local mode
 kubectl create namespace test-memcached
-operator-sdk test local ./test/e2e --up-local --namespace=test-memcached
+operator-sdk test local ./test/e2e --up-local --deploy-namespace=test-memcached
+kubectl delete namespace test-memcached
+
+# test operator in up local mode with --watch-namespace flag
+kubectl create namespace test-memcached
+operator-sdk test local ./test/e2e --up-local --deploy-namespace=test-memcached --watch-namespace=""
 kubectl delete namespace test-memcached
 
 # test operator in up local mode with kubeconfig
 kubectl create namespace test-memcached
-operator-sdk test local ./test/e2e --up-local --namespace=test-memcached --kubeconfig $KUBECONFIG
+operator-sdk test local ./test/e2e --up-local --deploy-namespace=test-memcached --kubeconfig $KUBECONFIG
 kubectl delete namespace test-memcached
 
 # test operator in no-setup mode
 kubectl create namespace test-memcached
 kubectl create -f deploy/crds/cache.example.com_memcacheds_crd.yaml
 # this runs after the popd at the end, so it needs the path from the project root
-trap_add 'kubectl delete -f test/test-framework/deploy/crds/cache.example.com_memcacheds_crd.yaml' EXIT
+trap_add 'kubectl delete --ignore-not-found -f test/test-framework/deploy/crds/cache.example.com_memcacheds_crd.yaml' EXIT
 kubectl create -f deploy/crds/cache.example.com_memcachedrs_crd.yaml
 # this runs after the popd at the end, so it needs the path from the project root
-trap_add 'kubectl delete -f test/test-framework/deploy/crds/cache.example.com_memcachedrs_crd.yaml' EXIT
+trap_add 'kubectl delete --ignore-not-found -f test/test-framework/deploy/crds/cache.example.com_memcachedrs_crd.yaml' EXIT
 kubectl create -f deploy/service_account.yaml --namespace test-memcached
 kubectl create -f deploy/role.yaml --namespace test-memcached
 kubectl create -f deploy/role_binding.yaml --namespace test-memcached
 kubectl create -f deploy/operator.yaml --namespace test-memcached
-operator-sdk test local ./test/e2e --namespace=test-memcached --no-setup
+operator-sdk test local ./test/e2e --deploy-namespace=test-memcached --no-setup
 kubectl delete namespace test-memcached
 popd